Remove assertionFromAuthenticator from StartAuthenticator.js and move to authenticatorAssertion
